Added if statement in front of mount

+++ system-scripts/mount_ram_drive.sh
@@ -1,7 +1,8 @@
 #!/bin/bash
-
-mkdir ~/org
-mkfs -q /dev/ram1 2048
-mount /dev/ram1 /home/pi/org
+if [ ! -d "/home/pi/org" ]; then
+mkdir /home/pi/org
+mkfs -q  /dev/ram1 2048
+mount -y /dev/ram1 /home/pi/org
 chown pi /home/pi/org
 chown pi /dev/ram1
+fi
